+++ tests/slo/internal/config/help.go
@@ -1,4 +1,4 @@
-package configs
+package config
 
 const (
 	mainHelp = `Commands:
